Extract ItemsInFolder + useAdminPublications instead of useAdminPublicationChildren in project folder settings
