+++ front/app/modules/commercial/project_folders/admin/containers/projects/index.tsx
@@ -8,30 +8,21 @@ import { updateProjectFolderMembership } from '../../../services/projects';
 import { isAdmin } from 'services/permissions/roles';
 
 // hooks
-import useAdminPublications, {
-  IAdminPublicationContent,
-} from 'hooks/useAdminPublications';
-import useAdminPublicationChildren from 'hooks/useAdminPublicationChildren';
+import useAdminPublications from 'hooks/useAdminPublications';
 import useAuthUser from 'hooks/useAuthUser';
-import useProjectFolder from '../../../hooks/useProjectFolder';
 
 // localisation
 import { FormattedMessage } from 'utils/cl-intl';
 import messages from '../messages';
 
 // components
-import {
-  List,
-  Row,
-  SortableList,
-  SortableRow,
-} from 'components/admin/ResourceList';
+import { List, Row } from 'components/admin/ResourceList';
 import { HeaderTitle } from 'containers/Admin/projects/all/StyledComponents';
 import ProjectRow from 'containers/Admin/projects/components/ProjectRow';
+import ItemsInFolder from './ItemsInFolder';
 
 // style
 import styled from 'styled-components';
-import { reorderAdminPublication } from 'services/adminPublications';
 
 const Container = styled.div`
   min-height: 60vh;
@@ -76,37 +67,17 @@ const AdminFolderProjectsList = ({
   });
 
   const authUser = useAuthUser();
-  const projectFolder = useProjectFolder({ projectFolderId });
-  const projectsInFolder = useAdminPublicationChildren({
-    publicationId: !isNilOrError(projectFolder)
-      ? projectFolder.relationships.admin_publication.data?.id
-      : undefined,
-    publicationStatusFilter: publicationStatuses,
-  });
-
   const [processing, setProcessing] = useState<string[]>([]);
 
-  if (isNilOrError(projectsInFolder) || isNilOrError(authUser)) {
+  if (isNilOrError(authUser)) {
     return null;
   }
-
-  const handleReorder = (itemId, newOrder) => {
-    reorderAdminPublication(itemId, newOrder);
-  };
 
   const addProjectToFolder = (projectFolderId: string) => (
     projectId: string
   ) => async () => {
     setProcessing([...processing, projectId]);
     await updateProjectFolderMembership(projectId, projectFolderId);
-    setProcessing(processing.filter((item) => projectId !== item));
-  };
-
-  const removeProjectFromFolder = (projectFolderId: string) => (
-    projectId: string
-  ) => async () => {
-    setProcessing([...processing, projectId]);
-    await updateProjectFolderMembership(projectId, null, projectFolderId);
     setProcessing(processing.filter((item) => projectId !== item));
   };
 
@@ -129,61 +100,7 @@ const AdminFolderProjectsList = ({
           <Spacer />
         </ListHeader>
 
-        {!isNilOrError(projectsInFolder) && projectsInFolder.length > 0 ? (
-          <SortableList
-            key={`IN_FOLDER_LIST${projectsInFolder.length}`}
-            items={projectsInFolder}
-            onReorder={handleReorder}
-            className="projects-list e2e-admin-folder-projects-list"
-            id="e2e-admin-folders-projects-list"
-          >
-            {({ itemsList, handleDragRow, handleDropRow }) => (
-              <>
-                {itemsList.map(
-                  (adminPublication: IAdminPublicationContent, index) => {
-                    return (
-                      <SortableRow
-                        key={adminPublication.id}
-                        id={adminPublication.id}
-                        index={index}
-                        moveRow={handleDragRow}
-                        dropRow={handleDropRow}
-                        lastItem={index === itemsList.length - 1}
-                      >
-                        <ProjectRow
-                          publication={adminPublication}
-                          actions={
-                            userIsAdmin
-                              ? [
-                                  {
-                                    buttonContent: (
-                                      <FormattedMessage
-                                        {...messages.removeFromFolder}
-                                      />
-                                    ),
-                                    handler: removeProjectFromFolder(
-                                      projectFolderId
-                                    ),
-                                    icon: 'remove',
-                                    processing: processing.includes(
-                                      adminPublication.publicationId
-                                    ),
-                                  },
-                                  'manage',
-                                ]
-                              : ['manage']
-                          }
-                        />
-                      </SortableRow>
-                    );
-                  }
-                )}
-              </>
-            )}
-          </SortableList>
-        ) : (
-          <FormattedMessage {...messages.folderEmptyGoBackToAdd} />
-        )}
+        <ItemsInFolder projectFolderId={projectFolderId} />
 
         {userIsAdmin && (
           <>
